Kotlin support
Workaround for Kotlin. It gets the first char of a class (first index);
it checks if before the class name there's another char, if true skip
this line; calculates the used lenght, adding to the first index, the
length of class name, plus a special char (like . (  ). At the end it
checks if the substring matches with a valid usage of class name, so
replace it with the URL.

PROS: it works.
CONS: if exists a string with " Main(", the workaround detects it like a
code line.